Move WarningModal component to higher level for new architecture change.
