+++ js/src/setup-mc/setup-stepper/setup-accounts/google-mc-account/warning-modal/index.scss
